update fundamental answers

+++ D6823-FYP_CSP-AE Fundamental-JKT/Problem C: Jojo and Arithmetic Progression/answer.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int N, M;
-    scanf("%d %d", &N, &M);
+    unsigned short N, M;
+    scanf("%hu %hu", &N, &M);
     for (int i = 0; i < M + 1; i++) printf("%d\n", N + i);
     return 0;
 }
